+++ test/production/app-dir/client-components-tree-shaking/app/client-import-namespace/client-module/client-module.js
